Migrate CycleMapGraph
